Use double for percent complete.

+++ test/src/module/command/backupTest.c
@@ -1851,7 +1851,7 @@ testRun(void)
         ProtocolParallelJob *job = protocolParallelJobNew(VARSTRDEF("key"), protocolCommandNew(strIdFromZ("x")));
         protocolParallelJobErrorSet(job, errorTypeCode(&AssertError), STRDEF("error message"));
 
-        int currentPercentComplete = 0;
+        double currentPercentComplete = 0;
         TEST_ERROR(
             backupJobResult((Manifest *)1, NULL, storageTest, strLstNew(), job, false, 0, NULL, &currentPercentComplete),
             AssertError, "error message");
